Replace old index.jsx
NOTE:
- Should buttons use onClick? is there an onSubmit?
